Split list_net/tests into multiple files.
