+++ xayn-ai/src/ltr/list_net/ndutils/io.rs
@@ -297,6 +297,8 @@ impl<'a> BinParamsWithScope<'a> {
 
 #[cfg(test)]
 mod tests {
+    use std::collections::HashSet;
+
     use ndarray::{arr1, arr2, Array1, Array2};
     use rand::{thread_rng, Rng};
 
@@ -385,5 +387,46 @@ mod tests {
                 assert_approx_eq!(f32, &fla1.data, &fla2.data, ulps = 0);
             }
         }
+    }
+
+    const EMPTY_BIN_PARAMS: &[u8] = &[0u8; 8];
+
+    const BIN_PARAMS_WITH_EMPTY_ARRAY_AND_KEY: &[u8] = &[
+        1u8, 0, 0, 0, 0, 0, 0, 0, // 1 entry
+        0, 0, 0, 0, 0, 0, 0, 0, // empty string key
+        1, 0, 0, 0, 0, 0, 0, 0, // 1 dimensional array
+        0, 0, 0, 0, 0, 0, 0, 0, // the dimension is 0
+        0, 0, 0, 0, 0, 0, 0, 0, // and we have 0 bytes of data data
+    ];
+
+    const BIN_PARAMS_WITH_SOME_KEYS: &[u8] = &[
+        2u8, 0, 0, 0, 0, 0, 0, 0, // 2 entries
+        3, 0, 0, 0, 0, 0, 0, 0, b'f', b'o', b'o', // key 1
+        1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // shape [0]
+        0, 0, 0, 0, 0, 0, 0, 0, // and data
+        3, 0, 0, 0, 0, 0, 0, 0, b'b', b'a', b'r', // key 2
+        1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, // shape [0]
+        0, 0, 0, 0, 0, 0, 0, 0, // and data
+    ];
+
+    #[test]
+    fn test_bin_params_is_empty() {
+        let params = BinParams::deserialize_from(EMPTY_BIN_PARAMS).unwrap();
+        assert!(params.is_empty());
+
+        let mut params = BinParams::deserialize_from(BIN_PARAMS_WITH_EMPTY_ARRAY_AND_KEY).unwrap();
+        assert!(!params.is_empty());
+
+        let array: Array<f32, IxDyn> = params.take("").unwrap();
+        assert_eq!(array.shape(), &[0]);
+    }
+
+    #[test]
+    fn test_bin_params_keys() {
+        let params = BinParams::deserialize_from(BIN_PARAMS_WITH_SOME_KEYS).unwrap();
+        let mut expected = HashSet::default();
+        expected.insert("foo");
+        expected.insert("bar");
+        assert_eq!(params.keys().collect::<HashSet<_>>(), expected);
     }
 }
